consolidate package controllers
